+++ frontend/src/components/layout/footer.tsx
