+++ frontend/src/components/questiontypes/imageupload_questiontype.tsx
@@ -14,7 +14,7 @@ import { useAppDispatch, useAppSelector } from "@/store/store";
 import { downloadFile } from "@/utils/helpers";
 
 import QuestionTypes, { DefaultQuestionTypeProps } from "./questiontypes";
-import { AwaitingChild } from "../awaiting";
+import { AwaitingChild } from "../layout/awaiting";
 import { SubmitButton } from "../submitButton";
 
 export interface ImageUploadQuestionTypeProps extends DefaultQuestionTypeProps {
