+++ frontend/src/components/questiontypes/longtext_questiontype.tsx
@@ -10,7 +10,7 @@ import { UpdateAnswer } from "@/store/slices/answerSlice";
 import { useAppDispatch, useAppSelector } from "@/store/store";
 
 import QuestionTypes, { DefaultQuestionTypeProps } from "./questiontypes";
-import { AwaitingChild } from "../awaiting";
+import { AwaitingChild } from "../layout/awaiting";
 import Easteregg from "../easteregg";
 
 export interface LongTextQuestionTypeProps extends DefaultQuestionTypeProps {
